Extracts ActivityLifecycleObserver class, #26.
